+++ lib/features/authenticated/loan_payment/presentation/pages/bloc/loan_payment_event.dart
@@ -11,8 +11,8 @@ class FetchLoanPayment extends LoanPaymentEvent {
   final String loanNumber;
   final int interestDays;
   final double interestRate;
-  final double loanBalance;
-  final double loanRefundAmount;
+  final int loanBalance;
+  final int loanRefundAmount;
   final String moduleCode;
   final String? issuedDate;
   final String? lastPaidDate;
